Move voice take into its own module
